Separating composables and activities in the exercises for Chapters 10, 14, and 15
