+++ Chapter10/Exercise10.01/app/src/main/java/com/example/popularmovies/DetailsActivity.kt
@@ -4,30 +4,7 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.filled.ArrowBack
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
-import androidx.compose.material3.TopAppBar
-import androidx.compose.material3.TopAppBarDefaults.topAppBarColors
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.style.TextOverflow
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import coil.compose.AsyncImage
+import com.example.popularmovies.ui.DetailsScreen
 import com.example.popularmovies.ui.theme.PopularMoviesTheme
 
 class DetailsActivity : ComponentActivity() {
@@ -40,100 +17,22 @@ class DetailsActivity : ComponentActivity() {
         const val EXTRA_POSTER = "poster"
     }
 
-    @OptIn(ExperimentalMaterial3Api::class)
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
         setContent {
             PopularMoviesTheme {
-                Scaffold(modifier = Modifier.fillMaxSize(),
-                    topBar = {
-                        TopAppBar(
-                            title = {
-                                Text(
-                                    text = stringResource(id = R.string.app_name),
-                                    color = MaterialTheme.colorScheme.onPrimary
-                                )
-                            },
-                            colors = topAppBarColors(MaterialTheme.colorScheme.primary),
-                            navigationIcon = {
-                                IconButton(onClick = { finish() }) {
-                                    Icon(
-                                        imageVector = Icons.AutoMirrored.Filled.ArrowBack,
-                                        contentDescription = stringResource(id = R.string.back),
-                                        tint = MaterialTheme.colorScheme.onPrimary
-                                    )
-                                }
-                            }
-                        )
-                    }) { innerPadding ->
-                    val extras = intent.extras
-                    val posterPath = extras?.getString(EXTRA_POSTER).orEmpty()
-                    DetailsView(
-                        title = extras?.getString(EXTRA_TITLE).orEmpty(),
-                        release = extras?.getString(EXTRA_RELEASE).orEmpty().take(4),
-                        overview = extras?.getString(EXTRA_OVERVIEW).orEmpty(),
-                        image = "$IMAGE_URL$posterPath",
-                        modifier = Modifier.padding(innerPadding)
-                    )
-                }
-            }
-        }
-    }
-}
+                val extras = intent.extras
+                val posterPath = extras?.getString(EXTRA_POSTER).orEmpty()
 
-@Composable
-fun DetailsView(
-    title: String,
-    release: String,
-    overview: String,
-    image: String,
-    modifier: Modifier = Modifier
-) {
-    Column(
-        modifier = modifier
-            .fillMaxWidth()
-            .padding(16.dp)
-    ) {
-        Row {
-            AsyncImage(
-                model = image,
-                contentDescription = title,
-                contentScale = ContentScale.Fit,
-                modifier = Modifier.size(160.dp)
-            )
-            Column(modifier = Modifier.padding(start = 16.dp)) {
-                Text(
-                    text = title,
-                    overflow = TextOverflow.Ellipsis,
-                    modifier = Modifier.fillMaxWidth()
-                )
-                Text(
-                    text = release,
-                    overflow = TextOverflow.Ellipsis,
-                    modifier = Modifier.fillMaxWidth()
+                DetailsScreen(
+                    title = extras?.getString(EXTRA_TITLE).orEmpty(),
+                    release = extras?.getString(EXTRA_RELEASE).orEmpty().take(4),
+                    overview = extras?.getString(EXTRA_OVERVIEW).orEmpty(),
+                    image = "$IMAGE_URL$posterPath",
+                    onBackButtonClick = { finish() }
                 )
             }
         }
-        Text(
-            text = overview,
-            overflow = TextOverflow.Ellipsis,
-            modifier = Modifier
-                .padding(top = 8.dp)
-                .fillMaxWidth()
-        )
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-fun DetailsViewPreview() {
-    PopularMoviesTheme {
-        DetailsView(
-            title = "Movie Title",
-            release = "2024-03-02",
-            overview = "Movie overview line 1\nLine 2",
-            image = "https://image.tmdb.org/t/p/w500/1E5baAaEse26fej7uHcjOgEE2t2.jpg"
-        )
     }
 }
